init; public release snapshot
This rendition of the project is the second iteration, rebuilt from scratch in Deno Fresh starting in early June 2023. The previous rendition was a Nuxt/Vue project build in early 2022 and was significantly less fleshed-out. Much of the content was written as far back as early 2020.

This git history will be re-written imminently to better reflect the process of building this site. It is being released in this flattened state merely for ease of publication.